Move `LibraryManga` to domain layer (#8126)

+++ app/src/main/java/eu/kanade/presentation/library/components/LibraryGridCover.kt
@@ -56,8 +56,8 @@ fun MangaGridCover(
 fun LibraryGridCover(
     modifier: Modifier = Modifier,
     mangaCover: eu.kanade.domain.manga.model.MangaCover,
-    downloadCount: Int,
-    unreadCount: Int,
+    downloadCount: Long,
+    unreadCount: Long,
     isLocal: Boolean,
     language: String,
     content: @Composable BoxScope.() -> Unit = {},
